updated some components and fixed balance issue
