+++ src/components/organisms/TradeDashboard/TradeChartArea/TradeChartArea.js
@@ -1,16 +1,19 @@
 import React from "react";
+import { useSelector } from "react-redux";
 import styled from "@xstyled/styled-components";
 import { TradeChart } from "components";
+import { marketInfoSelector } from "lib/store/features/api/apiSlice";
 
 const StyledTradeChart = styled.section`
   display: flex;
   grid-area: chart;
 `;
 
-export default function TradeChartArea(props) {
+export default function TradeChartArea() {
+  const marketInfo = useSelector(marketInfoSelector);
   return (
     <StyledTradeChart>
-      <TradeChart marketInfo={props.marketInfo} />
+      <TradeChart marketInfo={marketInfo} />
     </StyledTradeChart>
   );
 }
